+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] - 2022-##-##
+### Added
+- rating button tips
+
+### Changed
+- show type icon styling
 
 ## [1.13.0] - 2022-02-07
 ### Added